refactor: ShopTitle 경로 수정 및  지역변수로 수정함에 따라 카멜케이스로 수정
